Some cosmetic changes. Lets test it! ;)

+++ Radix_Tree/lib/radix_tree.cpp
@@ -23,7 +23,7 @@ const T & RadixTree::radix_tree<T,R>::get(const std::string & key) {
 template <class T, size_t R>
 RadixTree::node_ptr<T,R> RadixTree::radix_tree<T,R>::get(RadixTree::node_ptr<T,R> & n,
 							 const std::string & key,
-							 unsigned int d) {
+							 size_t d) {
 	if (n != nullptr) {
 		if (key.size() < d + n->path.size()) {
 			return nullptr;
@@ -62,7 +62,7 @@ template <class T, size_t R>
 RadixTree::node_ptr<T,R> RadixTree::radix_tree<T,R>::put(RadixTree::node_ptr<T,R> n,
 							 const std::string & key,
 							 const T & value,
-							 unsigned int d,
+							 size_t d,
 							 bool & created) {
 	if (n != nullptr) {
 		if (key.size() < d + n->path.size()) {
@@ -119,7 +119,7 @@ RadixTree::node_ptr<T,R> RadixTree::radix_tree<T,R>::put(RadixTree::node_ptr<T,R
 template <class T, size_t R>
 void RadixTree::radix_tree<T,R>::split(RadixTree::node_ptr<T,R> & n,
 				       const T & value,
-				       unsigned int p) {
+				       size_t p) {
 	std::string upper_path = n->path.substr(0, p);
 	std::string lower_path = n->path.substr(p, n->path.size()-p);
 	n->sons[lower_path[0]] = node_ptr<T,R>(new node<T,R>(lower_path, n->value));
@@ -133,10 +133,10 @@ template <class T, size_t R>
 void RadixTree::radix_tree<T,R>::find_diff_and_split(RadixTree::node_ptr<T,R> & n,
 						     const std::string & key,
 						     const T & value) {
-	unsigned int p_s = n->path.size();
-	unsigned int k_s = key.size();
-	unsigned int len = p_s < k_s ? p_s : k_s; // Take the min of both sizes
-	for (unsigned int i = 0; i < len; ++i) {
+	size_t p_s = n->path.size();
+	size_t k_s = key.size();
+	size_t len = p_s < k_s ? p_s : k_s; // Take the min of both sizes
+	for (size_t i = 0; i < len; ++i) {
 		if (n->path[i] != key[i]){
 			split(n, T(), i);
 			std::string branch = key.substr(i, key.size()-i);
@@ -150,7 +150,7 @@ void RadixTree::radix_tree<T,R>::find_diff_and_split(RadixTree::node_ptr<T,R> & 
 
 template <class T, size_t R>
 void RadixTree::radix_tree<T,R>::clean(RadixTree::node_ptr<T,R> n) {
-	for (unsigned int i = 0; i < n->r; ++i) {
+	for (size_t i = 0; i < n->r; ++i) {
 		if (n->sons[i] != nullptr) clean(std::move(n->sons[i]));
 	}
 	n.reset();
@@ -167,7 +167,7 @@ bool RadixTree::radix_tree<T,R>::contains(const std::string & key) {
 template <class T, size_t R>
 bool RadixTree::radix_tree<T,R>::contains(RadixTree::node_ptr<T,R> & n,
 					  const std::string & key,
-					  unsigned int d) {
+					  size_t d) {
 	if (n != nullptr) {
 		if (key.size() < d + n->path.size()) {
 			return false;
@@ -204,7 +204,7 @@ void RadixTree::radix_tree<T,R>::remove(const std::string & key) {
 template <class T, size_t R>
 bool RadixTree::radix_tree<T,R>::remove(RadixTree::node_ptr<T,R> & n,
 					const std::string & key,
-					unsigned int d,
+					size_t d,
 					bool & decrease) {
 	if (n != nullptr) {
 		if (key.size() < d + n->path.size()) {
@@ -250,7 +250,7 @@ bool RadixTree::radix_tree<T,R>::remove(RadixTree::node_ptr<T,R> & n,
 template <class T, size_t R>
 void RadixTree::radix_tree<T,R>::merge_with_only_son(RadixTree::node_ptr<T,R> & n) {
 	if (n->s == 1){
-		for (unsigned int i = 0; i < r; ++i) {
+		for (size_t i = 0; i < r; ++i) {
 			if (n->sons[i] != nullptr) {
 				merge(n, n->sons[i]);
 				break;
@@ -266,8 +266,8 @@ void RadixTree::radix_tree<T,R>::merge(RadixTree::node_ptr<T,R> & father,
 	father->s = son->s;
 	father->path += son->path;
 	if (father->s != 0) {
-	        unsigned int left = 0;
-	        for(unsigned int i = 0; i < r && left < father->s; ++i) {
+	        size_t left = 0;
+	        for(size_t i = 0; i < r && left < father->s; ++i) {
 		        if (son->sons[i] != nullptr) {
 		                father->sons[i] = std::move(son->sons[i]);
 				++left;
@@ -281,7 +281,7 @@ std::vector<std::string> RadixTree::radix_tree<T,R>::get_keys(const std::string 
 	vec_ptr vec;
 	vec = vec_ptr(new std::vector<std::string>());
 	if (prefix == "") {
-		for (unsigned int i = 0; i < r; ++i) {
+		for (size_t i = 0; i < r; ++i) {
 			if (roots[i] != nullptr) {
 				gather_keys(roots[i], "", vec);
 			}
@@ -295,10 +295,10 @@ std::vector<std::string> RadixTree::radix_tree<T,R>::get_keys(const std::string 
 template <class T, size_t R>
 void RadixTree::radix_tree<T,R>::get_keys(RadixTree::node_ptr<T,R> & n,
 					  std::string prefix,
-					  unsigned int d,
+					  size_t d,
 					  vec_ptr & v) {
 	if (prefix.size() < d + n->path.size()) {
-		for (unsigned int i = d; i < prefix.size(); ++i){
+		for (size_t i = d; i < prefix.size(); ++i){
 			if (prefix[i] != n->path[i-d])
 				return;
 		}
@@ -326,9 +326,9 @@ void RadixTree::radix_tree<T,R>::get_keys(RadixTree::node_ptr<T,R> & n,
 }
 
 template <class T, size_t R>
-std::string RadixTree::radix_tree<T,R>::get_prefix_cut(std::string & prefix, unsigned int d) {
+std::string RadixTree::radix_tree<T,R>::get_prefix_cut(std::string & prefix, size_t d) {
 	std::string cut = "";
-	for (unsigned int i = 0; i < d; ++i) {
+	for (size_t i = 0; i < d; ++i) {
 		cut += prefix[i];
 	}
 	return cut;
@@ -342,7 +342,7 @@ void RadixTree::radix_tree<T,R>::gather_keys(RadixTree::node_ptr<T,R> & n,
 		v->push_back(prefix + n->path);
 	}
 	if (n->s != 0) {
-		for (unsigned int i = 0; i < n->r; ++i) {
+		for (size_t i = 0; i < n->r; ++i) {
 			if (n->sons[i] != nullptr) {
 				gather_keys(n->sons[i], prefix + n->path, v);
 			}
@@ -355,9 +355,9 @@ void RadixTree::radix_tree<T,R>::show() {
 	std::cout << "digraph graphName{" << std::endl;
 	std::cout << "node [shape=record];" << std::endl;
 	// Node labels
-	int label = 0;
+	size_t label = 0;
 	std::cout << label << " [shape=record,label=\"{Root}\"];" << std::endl;;
-	for (unsigned int i = 0; i < r; ++i) {
+	for (size_t i = 0; i < r; ++i) {
 	        if (roots[i] != nullptr) {
 			++label;
 			show_label(roots[i], label);
@@ -365,7 +365,7 @@ void RadixTree::radix_tree<T,R>::show() {
 	}
 	// Node hierarchy
 	label = 0;
-	for (unsigned int i = 0; i < r; ++i) {
+	for (size_t i = 0; i < r; ++i) {
 	        if (roots[i] != nullptr) {
                         std::cout << 0 << "->";
 			++label;
@@ -377,7 +377,7 @@ void RadixTree::radix_tree<T,R>::show() {
 
 template <class T, size_t R>
 void RadixTree::radix_tree<T,R>::show_label(RadixTree::node_ptr<T,R> & n,
-					    int & label) {
+					    size_t & label) {
          std::cout << label << " [shape=record,label=\"{ <data> " << n->path;
 	 if (n->value != T())
 	         std::cout << " | <value> " << n->value;
@@ -385,7 +385,7 @@ void RadixTree::radix_tree<T,R>::show_label(RadixTree::node_ptr<T,R> & n,
 	 if (n->value != T())
 	         std::cout << "color=\"blue\"";
 	 std:: cout << "];" << std::endl;
-         for (unsigned int i = 0; i < n->r; ++i) {
+         for (size_t i = 0; i < n->r; ++i) {
 	        if (n->sons[i] != nullptr) {
 		        ++label;
 			show_label(n->sons[i], label);
@@ -395,10 +395,10 @@ void RadixTree::radix_tree<T,R>::show_label(RadixTree::node_ptr<T,R> & n,
 
 template <class T, size_t R>
 void RadixTree::radix_tree<T,R>::show(RadixTree::node_ptr<T,R> & n,
-				      int & label) {
+				      size_t & label) {
          std::cout << label << std::endl;
-	 int copy_label = label;
-         for (unsigned int i = 0; i < n->r; ++i) {
+	 size_t copy_label = label;
+         for (size_t i = 0; i < n->r; ++i) {
 	        if (n->sons[i] != nullptr) {
 		        std::cout << copy_label << "->";
 			++label;
